Support processing of embedded structs.
envconfig currently ignores embedded structs in Process. Embedding can
be helpful if you want to join together configuration from multiple
application subpackages and process in a single shot.

This updates envconfig to offer simple support for embedded structs.
Note that it completely ignores the name and tags on the struct.