Added function to detect unused configs
As service keeps growing, it starts to have tens of configurations, and
some of them become deprecated by the code but still persist in the
configuration.

In order to detect those vars, we can use the information of the config
provided and the config that is being parsed by current code version.