allow maps of slices
Example: `map[string][]int`

This wasn't allowed before because the separator for map elements was
the same as the separator for slice elements.

Allow custom processors to be configured to handle this, but also
preserve the existing behavior.

The usage also had to change, as it is no longer just commas. For
this, use the actual values of the current processor to simplicity.